Add documentation

+++ src/audio/sequence.rs
@@ -1,13 +1,18 @@
+//! Tone sequencer.
+
 use crate::misc::SampleRate;
 
 use super::tone::Tone;
 
+/// A sequence of tones.
+/// Will continue to the next tone when the current one is finished.
 pub struct Sequence {
     tones: Vec<Tone>,
     index: usize,
 }
 
 impl Sequence {
+    /// Create a new empty sequence.
     pub fn new() -> Self {
         Self {
             tones: Vec::new(),
@@ -15,14 +20,18 @@ impl Sequence {
         }
     }
 
+    /// Add a tone to the sequence.
     pub fn _chain(mut self, tone: Tone) -> Self {
         self.tones.push(tone);
         self
     }
 
-    // Seq format:
-    // Freq;time(s)
-    // 440;1.2
+    /// Create a sequence from a string.
+    /// The format is as follows:
+    /// ```text
+    /// Freq;time(s)
+    /// 440;1.2
+    /// ```
     pub fn from_seq(seq: &str, sample_rate: SampleRate) -> Self {
         let mut tones = Vec::new();
 
@@ -31,7 +40,7 @@ impl Sequence {
             let freq = freq.parse::<f32>().unwrap();
             let time = time.parse::<f32>().unwrap();
             tones.push(
-                Tone::new(freq, sample_rate).duration((sample_rate.output as f32 * time) as u32),
+                Tone::new(freq, sample_rate).duration((sample_rate.output as f32 * time) as usize),
             );
         }
